# docs/swagger.yaml
basePath: /
definitions:
  handlers.bulkError:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  handlers.bulkGetItemResult:
    properties:
      cas:
        type: string
      content: {}
      content_type:
        type: string
      error:
        $ref: '#/definitions/handlers.bulkError'
      id:
        type: string
      ok:
        type: boolean
    type: object
  handlers.bulkGetRequest:
    properties:
      collection:
        type: string
      ids:
        items:
          type: string
        type: array
      scope:
        type: string
    type: object
  handlers.bulkGetResponse:
    properties:
      bucket:
        type: string
      collection:
        type: string
      results:
        items:
          $ref: '#/definitions/handlers.bulkGetItemResult'
        type: array
      scope:
        type: string
    type: object
  handlers.bulkUpsertItemResult:
    properties:
      cas:
        type: string
      error:
        $ref: '#/definitions/handlers.bulkError'
      id:
        type: string
      ok:
        type: boolean
    type: object
  handlers.bulkUpsertRequest:
    type: object
  handlers.bulkUpsertResponse:
    properties:
      bucket:
        type: string
      collection:
        type: string
      results:
        items:
          $ref: '#/definitions/handlers.bulkUpsertItemResult'
        type: array
      scope:
        type: string
    type: object
...
info:
  contact: {}
  description: REST API with Couchbase (gocb v2), chi router, zap logging.
  title: Couchbase Go API
  version: "1.0"
paths:
...
  /buckets/{bucket}/docs/_bulk_get:
    post:
      consumes:
      - application/json
      description: Best-effort fetch multiple documents; per-item result included.
        200 even with partial failures.
      parameters:
      - description: Bucket name
        in: path
        name: bucket
        required: true
        type: string
      - description: Scope name (default _default)
        in: query
        name: scope
        type: string
      - description: Collection name (default _default)
        in: query
        name: collection
        type: string
      - description: Bulk get request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.bulkGetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.bulkGetResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties: true
            type: object
      summary: Bulk get documents
  /buckets/{bucket}/docs/_bulk_upsert:
    post:
      consumes:
      - application/json
      description: Best-effort upsert multiple documents; per-item result included.
        200 even with partial failures.
      parameters:
      - description: Bucket name
        in: path
        name: bucket
        required: true
        type: string
      - description: Scope name (default _default)
        in: query
        name: scope
        type: string
      - description: Collection name (default _default)
        in: query
        name: collection
        type: string
      - description: Bulk upsert request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.bulkUpsertRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.bulkUpsertResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties: true
            type: object
      summary: Bulk upsert documents
...
swagger: "2.0"